Fix model filtering based on range
We currently discard all models that are not in the range
  min_intercept < model.intercept_ < max_intercept

However, when we have more than min_samples_required but less than
history_window_size points in history, we force the intercept to 0
(fit_intercept=False).

This means that we actually discard all these models, until the history
contains history_window_size points.