Config: Add support for env
This adds support for configuring some key parameter using environment
variables. This is useful, and recommended (cf. twelve-factor app), when
running in a environment where changing
the configuration file across deployments (e.g. k8s in our case).

At the moment, only a subset of configuration parameters are supported.
It's currently quite a hack and should probably be merged with the
existing configuration mechanism, which currently supports file and cli
arguments.

Signed-off-by: Pierre Rust <pierre.rust@orange.com>