build(Dockerfile): Use PowerAPI global virtualenv
Using a user virtualenv prevents the container from being run as another
user with the `--user` docker/podman run flag.
Since PowerAPI v2.8.0, a global virtualenv available in the image, which
fixes this problem.

This is useful to run SmartWatts with a rootless container runtime
and have the output files with the same permissions as the running user.